Add basic Vagrantfile and an Ansible playbook
The idea here was to create a simple environment to speed up
development. Vagrantbox won't be the same or as a clean as native
base installation, but Vagrant environments are easier to throw
away and build up again.

I'll use Ansible to setup/install things on my machine, so I also
made a very simple Ansible runtime environment.